+++ granular_configuration_language/_json.py
@@ -7,7 +7,8 @@ from datetime import date, datetime
 from functools import partial, update_wrapper
 from uuid import UUID
 
-from granular_configuration_language import Configuration
+from granular_configuration_language import Configuration, LazyLoadConfiguration
+from granular_configuration_language._lazy_load_configuration import SafeConfigurationProxy
 
 
 def get_name(value: tabc.Callable) -> str:
@@ -43,6 +44,10 @@ def json_default(value: typ.Any) -> typ.Any:
     """
 
     if isinstance(value, Configuration):
+        return value.as_dict()
+    elif isinstance(value, LazyLoadConfiguration):
+        return value.config.as_dict()
+    elif isinstance(value, SafeConfigurationProxy):
         return value.as_dict()
     elif isinstance(value, UUID):
         return str(value)